requirements.txt
anthropic>=0.18.0
fastapi>=0.109.0
pydantic>=2.6.0
python-dotenv>=1.0.0
uvicorn>=0.27.0
